requirements.txt
streamlit
openai-agents
langchain
langchain-openai
langchain-community
faiss-cpu
pdfplumber
python-dotenv
openai